fatmug_backend/requirement.txt
Django>=3.2
djangorestframework
aiofiles
asgiref==3.8.1
psycopg2-binary>=2.9,<3.0
django-cors-headers
channels